Generate new UUIDs
TextMate and other similar editors require the UUID value to be globally unique, I am not aware of VSCode's usage of UUIDs. For the Comments file the UUID was shared with the similar preference item from the JavaScript bundle causing the conflict. For the theme files the UUID conflict was with the language grammar files in this bundle.