Use 'contentName' instead of 'name' to mark embedded jsx expressions
Jsx expressions currently use `name` to mark the entire expresion (including the opening and closing braces) as `meta.embedded.expression`. This causes VS Code to treat the entire expresssion as JavaScript for basic language commands such as highlighting.

This fix instead changes jsx expressions to only mark the expression itself (and not the braces) as `meta.embedded.expression`. This lets VS Code correctly handle commenting in a few edge cases